refactor and add __all__ to __init__.py
